// PIA projekat/frontend/frontend/src/app/models/user.ts
export class User{
  username: string = "";
  password: string = "";
  securityQuestion: string = "";
  securityAnswer: string = "";
  firstname: string = "";
  lastname: string = "";
  gender: string = "";
  address: string = "";
  phone: string = "";
  email: string = "";
  profilePicture: string = "";
  creditCardNumber: string = "";
  type: string = "gost";
  status: string = "pending";
  restaurant: string = "";
  numberOfNotComing: number = 0;
}